RowResultDecoder abstraction
